This commit contains all kinds of unfinished work which was likely meant to be split into multiple commits.
I had just finished working on the ba2d and camera files and started doing some integration when I figured out that
I would have to change the interfaces since it needs to me changed such that the depth is dynamic.

(See branch issues for other notes I made about this branch)

A small list of misc changes in this commit:
- Changed pieGetRescaling factor to float since all callers were dividing the result by 100.f
- Removed unimplemented sound function declarations.
- Disabled droid animations which are in the logical update
- Changed the lightmap pixel format to reduce upload time

If you try to use this branch, it'll be quite slow, this is because the depth of the camera is too large, causing algorithms to search further than necessary.

The interfaces I changed for the camera class might seem awkward now, but I did them before I found out about the depth issue.